avionic-design/medatom: Major updates.
This is pretty much the state that will be demoed at the Medica fair. It
is not completely cleaned up yet, but is working quite well.

One of the major features/changes is that the system is booted using
systemd instead of initscripts. Some of the services have no systemd
service files, though, so things like SSH and telnet daemons will not be
started by default.

Another big change is the move to NetworkManager for network device
configuration. This ties in neatly with the systemd boot process, so
that the potential long delay before an IP address can be obtain via
DHCP doesn't hog the rest of the boot process.

Some things that need cleaning up are the removal of the old initscripts
that are now no longer required.

As mentioned above, some more work is required to provide systemd
service files for all remaining services (SSH, telnet, etc).