avionic-design/medatom/rootfs: Start lldpd via systemd.
This commit adds a systemd socket unit for the lldpd control socket,
which will be passed to the lldpd daemon at startup. A systemd service
unit is also provided. Furthermore, this commit adds the _lldpd user and
group, without which the daemon will not start correctly.

This commit also adds a NetworkManager dispatcher script which will
(re)start the lldpd service via systemd whenever a network connection is
brought up to ensure that LLDP frames are sent.