avionic-design/medatom/rootfs: Skip root filesystem check on boot.
This commit adds the /fastboot file to the root filesystem, which is
used by systemd (as well as by other distributions) to check if the root
filesystem check should be skipped. Since the root filesystem is a
squashfs image, for which no fsck tool exists, it makes obviously no
sense to perform this check on boot.