avionic-design/medatom/rootfs: Fix GUI startup.
Only start the remote-control binary if a configuration file is present.
Otherwise, start awesome by default. remote-control runs "headless" by
default if no configuration file is available but a session running a
headless remote-control is useless. This commit assumes that if the
configuration file is missing, the device is booted in development mode
and remote-control will be run manually.