avionic-design/medatom: Add live USB creation script.
This script creates and formats partitions in a standard way used to
boot Medatom units from a USB drive. The extlinux bootloader is also
installed to the MBR of the device. No actual data files are copied,
however. At least the vmlinuz, initrd.gz and rootfs.img files are
required for the USB drive to boot properly.