medatom/rootfs: Load tm6000 and tm6000-alsa modules.
These modules support the Terratec Cinergy Hybrid (XE) USB sticks that
are used to receive analog TV and radio programs. Both modules are in
the staging tree and therefore are not loaded automatically when a
supported device is plugged. This commit also adds the binary firmware
required by the xc3028 tuner.