medatom/rootfs: Use ntpd in favour of ntpdate.
ntpdate has been officially deprecated. ntpd can be used as a drop-in
replacement using the -q (synchronize time and exit) option.

On devices with long uptimes, ntpd is the preferred choice, however
because it continually updates the clock to compensate for local clock
drift. Note that the "iburst" option should be used in the ntp.conf
configuration file to make sure that the time is synchronized quickly,
otherwise it may take a couple of minutes.