avionic-design/plutux: Re-generate hostname at bootup.
This will update the script which is responsible for updating the hostname. It
checks now if the hostname is already set and if so if the name still maches
the device mac.
It not we remount the persistent storage to make it writable and the
regenerate the hostname. Finally we make the persistent storage read-only
again.