avionic-design/plutux: Backup old Xsession.log
Since i had random restarts of the nodm.service and not output on debug serial
i decided to check the Xsession.log. But the file is trunceted on each session
start. So all the information is lost.
This will extend the session startup script to backup the Xsession.log of a
previous session to preserve the possible debug info. And to save some space
we use gzip.