medatom/rootfs: Set keyboard layout in config file.
Setting the keyboard layout by running setxkbmap in an Xsession startup
script has some disadvantages: sometimes the call will simply not work,
outputting something along the lines of "cannot open input method", and
the settings will be lost when a keyboard is attached after the
setxkbmap call.

Eventually the keyboard layout configuration should be moved to
persistent storage so it can be overridden by configuration tarballs.