platform/plutux: Reduce unnessessary xrandr calls.
In order to select the right screen output, we need to know which display is
connected. Therefor we call xrandr an parse the output. This causes screen
flicker and forces the xserver to reload all input drivers.

This patch will reduce the xrandr calls to avoid the flickering by storing the
output of the first xrandr call and work only on the output.