medatom/rootfs: Add ALSA configuration file.
The configuration file creates two virtual PCM devices which can be used
to capture the left or right channels separately, respectively. This is
required in some situations because otherwise ALSA will capture the left
and right channels as stereo signal and mix them down to a mono signal,
which will halve the volume if only one of the channels actually has a
signal.

On Medatom this is used by VoIP where the left channel is the normal
microphone for handset telephony and the right channel is the microphone
for hands-free calls.