medatom/rootfs: Fix initial calibration.
Due to some weird bug, the X server closes and reopens input devices
after the calibration, which causes the calibration data to be lost.
This can be worked around by exiting the X session after the calibration
which will cause a restart of the session and the calibration data being
loaded from the configuration file generated before.