medcom-wide: Fix initial calibration.
In the past our entire xorg.conf.d directory was a symlink to persist.
Now we have only a symlink to the calibration file. This means we can
not override the file, since the it symlink stored on a read-only
filesystem. So we need to make a readlink on the calibration file in
order to get the correct target for the copy operation.