plutux/rootfs: Replace xorg.conf.d symlink with static config.
In the beginning we made the the directory /etc/X11/xorg.conf.d a symlink to
/persist/etc/X11/xorg.conf.d. This did allow us to play with the Xorg config.
Now that we are know what we are doing, this is no longer needed.

So this commit will add a default configuration with generic input support.
It disables the LVDS output and select the HDMI output with disabled DPMS.

With the 90-undefined.conf we still have the possibility to override
configurations if needed.