medcom-wide/rootfs: Add Xsession.d hook for persistent storage.
With this commit the /etc/X11/Xsession.d directory can now be mirrored
on the persistent storage and its contents will be sourced before all
other scripts. This feature can be used for example to set environment
variables for the X session.
We also update the 50-determine-startup script to allow an overriding
the STARTUP environment variable.