medatom/rootfs: Add support for log file rotation
The two log files /var/log/messages and /var/log/remote-control.log tend
to grow rather large after some time, so they are rotated periodically
and old logs are compressed to take up less space. Currently files can
be a maximum of 16 MiB and if that limit is not reached they will be
rotated daily. Of each of these a maximum of 7 copies will be kept so
unless the logs are spammed a full week of logging output will remain on
the device.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Julian Scheel <julian@jusst.de>