medatom/rootfs: Use per-interface iflinkd instances
Instead of trying to handle all interfaces in the system with a single
instance of the iflinkd daemon, instantiate it per interface. This has
the big advantage that the configuration of one interface cannot be
blocked by the configuration of another. Furthermore it allows each
interface to be reconfigured separately.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Julian Scheel <julian@jusst.de>