medatom/rootfs: Support Medatom-Wide ALSA configuration
In order to support configuration of ALSA on multiple variants (like the
Medatom and Medatom Wide), some major restructuring needs to be done. A
new directory /etc/alsa is added, which contains state and configuration
files for each variant along with a small script which is run on boot to
perform configuration that cannot be encoded in either the state or the
configuration files.

A new script is provided by the util/medatom-utils package that loads
the ALSA state files depending on the variant and runs the corresponding
script. A corresponding systemd unit is also provided by the package.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Julian Scheel <julian@jusst.de>