plutux/rootfs: Use per-interface iflinkd instances.
Instead of handling all interfaces in the system with a single instance of
the iflinkd daemon, instantiate it per interface. This is now nessessary
because we no longer install the system wide service file.

It also has the big advantage that the configuration of one interface can't
be blocked by the configuration of another. And it configures only the
interfaces we know and trust. Furthermore it allows each interface to be
reconfigured separately.

Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-by: Julian Scheel <julian@jusst.de>