medcom-wide/rootfs: Add bridge setup for networking.
On medcom-wide gen2 we have a second network port, instead of a switch on
gen1, that needs to be configured. So to provide a similar behavior on gen1
and gen2 board, we use a bridge. This is less painful to configure (in the
field) than routing and performs very well. To prevent losing connection (a
thing that McAdmin can't handle), the bridge is created on boot-up and not
dynamically as we tried before.

Change-Id: Ifad132e33ddb8ec3e3eaf9517d2bf8b854fed623
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/175
Reviewed-by: Julian Scheel <julian.scheel@avionic-design.de>