plutux/rootfs: Specifiy files to be rotated.
In the past we have rotated all files in /var/log/ ending with .log.
This includes also the Xorg.0.log. But this file is not under syslog
control. So that we loose all new log messages, once the file has been
rotated.
Also we now inform the syslog daemon the way it is suppost to be done,
instead of restarting the service.

Change-Id: I91793c3938d43e75fedc84df64e8e1769a9c0077
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/163
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>