medcom-wide/rootfs: Set lldp link for remote-control.
This adds a new configuration file for remote-control to specify the
interface to be used by the remote-controls internal lldp-monitor. By
default the remote-control used the interfaces with the default route.
But because we now use bridges and lldp is not allowed there, we just
tell remote-control to use the primary interface.

Change-Id: Id83d282bfa5cb6b1940af18f35e5fc5b1da9a228
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/182
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>