medcom-wide/rootfs: Allow bridge configuration.
This allows us to override the bridge defaults if needed via configuration
file. Doing this we also change the default settings. First of we keep the
bridge in learning state for at least 5 sec. Second we create the bridge
with STP disabled.
The configuration file is included, so that you set the enviroment variables
directly. This sample shows the usage:

	$ cat /persist/etc/bridge.conf
	BRIDGE_FORWARD_TIME=5
	BRIDGE_STP=off

Change-Id: I8cf2bfbcd5d3dab9bf9a81b1b97ac432b74f5c1b
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/183
Reviewed-by: Julian Scheel <julian.scheel@avionic-design.de>