medcom-wide/rootfs: Add plug for handset to asound.conf
Add a new plug named handset to the asound.conf, so we can use this name
from inside an application. This will avoid the need of to much knowlege
of the os and its configuration inside an application. Otherwise we need
to enumerate all devices and select one specific wich then depends on
the hardware we are running at.

Change-Id: I1082636e4ed57d5158739b5a7a4892b492c33da6
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/387
Reviewed-by: Julian Scheel <julian.scheel@avionic-design.de>