medcom-wide/rootfs: Enable zram based swap on bootup.
With the back-ported zram implementation we have a quite stable implementation,
which allows us to enable zram based swap. This should improve the stability of
the system under high memory load conditions.

Together with the default start, we make the zram start depend on basic.target
instead of multi-user.target. We also use only one zram block device with a
fixed size, instead of depending on the number of the cpus and the actual
system memory.

Change-Id: I7b8a7d486d4013149273c8859e10ba9850784ee4
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/740
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>