tec/initrd: Update device naming script
Boards running with T30 should be named with -ng. The simplest way to do this,
is to check the number of CPU cores.
Wile adding this we notice, that the script does not set the hostname, when no
persistent hostname file is present. So we restructure the file a bit to solve
this issue.

Change-Id: I2907572feba871e40a0c3134e4aabff508c7ba9a
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/686
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>