medcom-wide/rootfs: Fix default package selection
In the past, we had a default package selecting as part of the platform. This
default selection is outdated for quite some time now. Remote-control and rum
have been moved to vendor/avionic-design and the virtual package selection is
not working as expected. To solve this, we move the important packages like
L4T, remote-control and rum at the top of the list and use the correct config
entries to run a terminal in default configuration. Finally we remove the
packages which are not essential to run.

Change-Id: I50654d3f3af12f4b0614c93a710bf8adb59bcfaa
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/816
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>