medcom-wide/rootfs: Fix linphone usb-handset usage
When using linphone in combination with the usb-handset, we need to make sure
we use the correct soundcard. In this case a dmix plug, in order to be able to
play additional tones while a call is running. Therefore we need to tell
linphone to create the non-HW soundcard called handset and to use this card
for playback.

Due to an issue (SIGFPE) within ALSA, we need to remove the sample-rate module
from the dmix plug.

Change-Id: I3368eef4b2752ba232bfc3654312cd156ff70a60
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/869
Reviewed-by: Julian Scheel <julian@jusst.de>