medcom-wide/rootfs: ucm: Don't mute lineout for handset
The handsets speaker is wired to the Line Out channel of the wm8903. Therefore
it is a bad idea to mute this one when entering Voice Call/Handset state.
Instead the volume should just be kept and the Line Out has to be enabled.
This probably went in by accident at all.

Change-Id: I600ac8d7c38d4b3ae2f134c7b17f0cebd7578389
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/1074
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>