medcom-wide/rootfs: ucm: HACK: Enable bypass in voip mode
Enabling the hardware bypass of the line in paths somehow works around a bug
where the audio codec does not deliver any data anymore. As it does no harm to
bypass the line, enable it for now. We should try to figure out the actual
cause of the problem though.

Change-Id: Id5b1c378a0c70bd66d64f1c2095a26d96f82c3e7
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/1316
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>