medatom/rootfs: Start dbus session bus
This is a port of commit f544faff0 from medcom-wide.

Start a dbus session bus along with the Xsession. Additionally store the
session bus into a status file and evaluate this in the main bash-profile.
This hooks every bash session into the dbus session bus, which is associated
to the currently running Xsession. Being part of the dbus session bus one can
start applications, like remote-control-browser, out of a ssh session without
loosing accesibility access.

Change-Id: Ief7fd272cd3a5bed148e5be77362f294e3a1a36f
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/4545
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>