medatom/rootfs: Add egalax touch configuration
The new medatom terminal 1469 no longer uses the sx8634 sensor bars. Instead
there is a big touch that also covers the sensor buttons. Therefore we have
to calibrate the touch to fit the display. The sensor buttons have to be
handled by the GUI by adding buttons at the last line - there is no key event
generated by the OS!

Signed-off-by: Dirk Leber <dirk.leber@avionic-design.de>
Change-Id: Ifc9cf826b6bc250038c5340ffde1d171f8915dfd
Reviewed-on: http://review.adnet.avionic-design.de/5390
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>