medcom-wide/make-sdcard: fix sfdisk partioning bug
The sfdisk tool used for partitioning SD-cards has, especially in an
older 2.23.2 version on CentOS 7, problems re-reading the partitions it
just created. As even the --force parameter does not prevent sfdisk from
grinding to a halt, use --no-reread instead. This avoids races with
udev, as stated in numerous bug reports:

https://bugs.launchpad.net/ubuntu/+source/util-linux/+bug/942788
https://github.com/coreos/bugs/issues/152
https://bugs.centos.org/view.php?id=986

Change-Id: I09c4a90c99e324abb8469d6bad1465713d7c8b32
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5446
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>