*/make-{device}: fix sfdisk usage
The sfdisk tool has since coreutils 2.26 the -u parameter dropped,
making it incompatible to previous versions for the scripts which depend
on it. Being annoyingly difficult to detect, this problem requires a bit
of hack to make both old and new sfdisk versions work. Apply it to
anything using sfdisk with devices sizes.

Change-Id: I16e9804cebd05bd0c2e2c7dfb18f25c68bbc8427
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5467
Reviewed-by: Hendrik Lippek <hendrik.lippek@avionic-design.de>