medcom-wide/setup: add make-sdcard variant
Add another incarnation of the make-sdcard script, fit specifically for
the initramfs environment it is to be run in. Includes a workaround for
a mkfs path issue and handles mmcblk partition names. Works with both
pre and post core-utils 2.27 sfdisk variants. Lacks colour as its output
will most likely end up in a log anyway.

Change-Id: Id154c8b2ff1d6a3fcb9ee6e10a3fe3fcff48ce01
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5495
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>