medatom/rootfs: recycle dbus session daemon
If the remote-control process dies for any reason, the X session is
restarted. As the dbus session daemon does not exit with this session
but is bound to the nodm service, the startup script spawned additional
dbus session daemons. Prevent this by probing for an existing daemon and
recycling the session if it works.

Change-Id: I8d96dd1579d60b11bda773bf239562c953a0484d
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5659
Reviewed-by: Julian Scheel <julian@jusst.de>