medcom-wide/rootfs/iflinkd: Split out setting the multicast route
Currently the multicast route is set when the bridge is setup. However
we will need a second bridge and the multicast route should not be set
on both. To fix this we split setting the multicast route in a
dedicated script that is the enabled for br0.

While at it also fix the path to the 'ip' utility, it is in /sbin and
not /bin.

Change-Id: I8fec9977998a53141a3a682ce9696794bcbc7e08
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6309
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>