medcom-wide/rootfs: Add support for guest VLAN
Allow putting the guest traffic in a dedicated VLAN. For this the
config file /persist/etc/vlan.conf should define GUEST_VLAN with the
VLAN ID to use. When the guest VLAN is set we then use a dedicated
bridge that link the guest VLAN interface with the guest port.

We also need to add an ebtable broute rule to avoid having the VLAN
tagged traffic delivered to the bridge instead of the proper VLAN
interface.

Change-Id: Ia0b5a47492045dade05eb3ba65afbde65c9ce1c7
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6311
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>