medcom-wide/rootfs: Remove start of zram/swap
Once in low memory situation, the zram makes it even worse - the system
is busy by swapping and compressing data making it unresponsive.
Especially while streaming, disabling the zram is way more stable.
On the other hand - with zram, we have more memory while browsing but
the system gets unresponsive too if swap is nearly full. So the best
solution is not to use it.

Change-Id: I77fbae84767558c613af1f8809f9010def6d37a3
Signed-off-by: Dirk Leber <dirk.leber@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6713
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>