medatom/rootfs: Set alsa, ucm and linphone config for 1496
Since usb handset was not supported in medatom, we adapt the config
from medcom. We don't need dynamic switching for linphone, since
the usb handset will be the only device supported for telephony.
Note that we have to use the card names because the order depends on
whether the handset is plugged during boot or not.

Change-Id: Ia8b931c7faa0d47208c8efb717b90928c642ace9
Signed-off-by: Dirk Leber <dirk.leber@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6843
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>