medcom: use one version to rule them all!
Seriously, having three similar OS on similar platforms need to have the
same version. Also, less of a hassle when incrementing.

The inclusion weirdness is a result of doing that relative to the
directory the current Makefile resides in. While being reliable, this
breaks when spaces occur somewhere in the path.

Change-Id: I664f8d5c3f8c06ece832b889c95a2e125766ce27
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7585
Reviewed-by: Meike Vocke <meike.vocke@avionic-design.de>